+++ src/main/java/racingcar/domain/CarMoveUtil.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.Random;
