+++ src/test/java/racingcar/domain/CarManagerTest.java
@@ -1,11 +1,6 @@
-package racingcar;
+package racingcar.domain;
 
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 
 import java.util.Arrays;
